refactor: MenuGroupService 분리
